soppa/1DT905 - Objektorienerad/a01-labs/lab4/task3/app/src/test/java/com/task3/WolfTest.java
package com.task3;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Test;

public class WolfTest {
    Wolf wolf = new Wolf();
    Robot robot = new Robot();
    

    @Test
    public void IconTest() {
        assertEquals("🐺", wolf.getIcon());
    }

    @Test
    public void movementTest() {

        Integer tempCol = wolf.getCurrentColumn();
        Integer tempRow = wolf.getCurrentRow();
        wolf.lostFight();
        wolf.wonFight();
        wolf.wolfMovement(robot, wolf);
        assertEquals(tempCol, wolf.getCurrentColumn());
        assertEquals(tempRow, wolf.getCurrentRow());
    }

}
